requirements.txt
asgiref==3.5.0
astor==0.8.1
attrs==21.4.0
bandit==1.7.2
cfgv==3.3.1
darglint==1.8.1
distlib==0.3.4
Django==3.2
django-debug-toolbar==3.2.4
django-split-settings==1.1.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
docutils==0.18.1
eradicate==2.0.0
Faker==13.3.0
filelock==3.6.0
flake8==3.9.2
flake8-bandit==2.1.2
flake8-broken-line==0.4.0
flake8-bugbear==21.11.29
flake8-commas==2.1.0
flake8-comprehensions==3.8.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-eradicate==1.2.0
flake8-isort==4.1.1
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.5
flake8-string-format==0.3.0
gitdb==4.0.9
GitPython==3.1.27
identify==2.4.11
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pbr==5.8.1
pep8-naming==0.11.1
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.15.0
psycopg2-binary==2.9
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.11.2
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
restructuredtext-lint==1.4.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stevedore==3.5.0
testfixtures==6.18.4
toml==0.10.2
tomli==2.0.1
types-pytz==2021.3.5
types-PyYAML==6.0.4
typing-extensions==3.10.0.2
virtualenv==20.13.2
wemake-python-styleguide==0.16.0
gunicorn
